+++ app/src/main/java/com/sine_x/material_wecenter/models/Dynamic.java
@@ -141,12 +141,12 @@ public class Dynamic {
     }
 
     public static class UserInfoEntity {
-        private long uid;
+        private int uid;
         private String user_name;
         private String signature;
         private String avatar_file;
 
-        public void setUid(long uid) {
+        public void setUid(int uid) {
             this.uid = uid;
         }
 
@@ -162,7 +162,7 @@ public class Dynamic {
             this.avatar_file = avatar_file;
         }
 
-        public long getUid() {
+        public int getUid() {
             return uid;
         }
