+++ app/src/main/java/com/sine_x/material_wecenter/models/ExploreItem.java
@@ -148,11 +148,11 @@ public class ExploreItem {
     }
 
     public static class UserInfoEntity {
-        private long uid;
+        private int uid;
         private String user_name;
         private String avatar_file;
 
-        public void setUid(long uid) {
+        public void setUid(int uid) {
             this.uid = uid;
         }
 
@@ -164,7 +164,7 @@ public class ExploreItem {
             this.avatar_file = avatar_file;
         }
 
-        public long getUid() {
+        public int getUid() {
             return uid;
         }
