+++ app/src/main/java/com/sine_x/material_wecenter/models/LoginProcess.java
@@ -2,15 +2,15 @@ package com.sine_x.material_wecenter.models;
 
 public class LoginProcess {
 
-    private long uid;
+    private int uid;
     private String user_name;
     private String avatar_file;
 
-    public long getUid() {
+    public int getUid() {
         return uid;
     }
 
-    public void setUid(long uid) {
+    public void setUid(int uid) {
         this.uid = uid;
     }
