+++ app/src/main/java/com/sine_x/material_wecenter/models/QuestionDetail.java
@@ -191,12 +191,12 @@ public class QuestionDetail {
         }
 
         public static class UserInfoEntity {
-            private long uid;
+            private int uid;
             private String signature;
             private String user_name;
             private String avatar_file;
 
-            public void setUid(long uid) {
+            public void setUid(int uid) {
                 this.uid = uid;
             }
 
@@ -216,7 +216,7 @@ public class QuestionDetail {
                 this.avatar_file = avatar_file;
             }
 
-            public long getUid() {
+            public int getUid() {
                 return uid;
             }
 
@@ -380,12 +380,12 @@ public class QuestionDetail {
         }
 
         public static class UserInfoEntity {
-            private long uid;
+            private int uid;
             private String signature;
             private String user_name;
             private String avatar_file;
 
-            public void setUid(long uid) {
+            public void setUid(int uid) {
                 this.uid = uid;
             }
 
@@ -405,7 +405,7 @@ public class QuestionDetail {
                 this.avatar_file = avatar_file;
             }
 
-            public long getUid() {
+            public int getUid() {
                 return uid;
             }
